Conflicts:
	src/common/cbasetypes.h
	src/login/lobby.cpp
	src/map/ai/ai_char_normal.cpp
	src/map/entities/charentity.cpp
	src/map/entities/charentity.h
	src/map/map.cpp
	src/map/map.h
	src/map/packet_system.cpp
	src/map/packets/menu_merit.cpp
	src/map/packets/party_member_update.cpp
	src/map/packets/party_member_update.h
	src/map/party.cpp
	src/map/party.h
	src/map/utils/charutils.cpp
	src/map/utils/charutils.h
	src/map/utils/fishingutils.cpp
	src/map/utils/zoneutils.cpp
	win32/DSGame-server/DSGame-server.vcxproj
	win32/DSGame-server/DSGame-server.vcxproj.filters

+++ src/map/packets/party_member_update.h
@@ -42,7 +42,7 @@ class CPartyMemberUpdatePacket : public CBasicPacket
 {
 public:
 
-	CPartyMemberUpdatePacket(CCharEntity* PChar, uint8 MemberNumber, uint8 ZoneID);
+	CPartyMemberUpdatePacket(CCharEntity* PChar, uint8 MemberNumber, uint16 ZoneID);
 	CPartyMemberUpdatePacket(uint32 id, const int8* name, uint16 memberFlags, uint8 ZoneID);
 };
