+++ src/map/party.cpp
@@ -182,7 +182,7 @@ void CParty::AssignPartyRole(int8* MemberName, uint8 role)
 *																		*
 ************************************************************************/
 
-uint8 CParty::MemberCount(uint8 ZoneID)
+uint8 CParty::MemberCount(uint16 ZoneID)
 {
 	uint8 count = 0;
 
@@ -945,7 +945,7 @@ void CParty::SetQuaterMaster(CBattleEntity* PEntity)
 *																		*
 ************************************************************************/
 
-void CParty::PushPacket(uint32 senderID, uint8 ZoneID, CBasicPacket* packet)
+void CParty::PushPacket(uint32 senderID, uint16 ZoneID, CBasicPacket* packet)
 {
 	for (uint32 i = 0; i < members.size(); ++i)
 	{
